refactor: rename component to camelCase format
