+++ src/circularProgressBase/styles.d.ts
