+++ src/circularProgressBase/styles.ts
