+++ src/circularProgressBase/types.d.ts
